requirements.txt
Django==5.1.7
django_ckeditor==6.7.2
django_taggit==6.1.0
shortuuid==1.0.13
